Create interface SAMLSSOServiceProviderDAORegistryImpl
