+++ components/identity-core/org.wso2.carbon.identity.core/src/main/java/org/wso2/carbon/identity/core/persistence/IdentityPersistenceManager.java
@@ -22,6 +22,7 @@ import org.wso2.carbon.identity.core.dao.OpenIDAdminDAO;
 import org.wso2.carbon.identity.core.dao.OpenIDUserDAO;
 import org.wso2.carbon.identity.core.dao.ParameterDAO;
 import org.wso2.carbon.identity.core.dao.SAMLSSOServiceProviderDAO;
+import org.wso2.carbon.identity.core.dao.SAMLSSOServiceProviderRegistryDAOImpl;
 import org.wso2.carbon.identity.core.dao.XMPPSettingsDAO;
 import org.wso2.carbon.identity.core.model.OpenIDAdminDO;
 import org.wso2.carbon.identity.core.model.OpenIDUserDO;
@@ -235,7 +236,7 @@ public class IdentityPersistenceManager {
      */
     public boolean addServiceProvider(Registry registry, SAMLSSOServiceProviderDO serviceProviderDO)
             throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDAO.addServiceProvider(serviceProviderDO);
     }
     /**
@@ -246,7 +247,7 @@ public class IdentityPersistenceManager {
      * @throws IdentityException
      */
     public SAMLSSOServiceProviderDO uploadServiceProvider(Registry registry, SAMLSSOServiceProviderDO samlssoServiceProviderDO) throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDAO.uploadServiceProvider(samlssoServiceProviderDO);
     }
 
@@ -258,23 +259,23 @@ public class IdentityPersistenceManager {
      */
     public SAMLSSOServiceProviderDO[] getServiceProviders(Registry registry)
             throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDOA = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDOA = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDOA.getServiceProviders();
     }
 
     public boolean removeServiceProvider(Registry registry, String issuer) throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDAO.removeServiceProvider(issuer);
     }
 
     public SAMLSSOServiceProviderDO getServiceProvider(Registry registry, String issuer)
             throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDAO.getServiceProvider(issuer);
     }
 
     public boolean isServiceProviderExists(Registry registry, String issuer) throws IdentityException {
-        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderDAO(registry);
+        SAMLSSOServiceProviderDAO serviceProviderDAO = new SAMLSSOServiceProviderRegistryDAOImpl(registry);
         return serviceProviderDAO.isServiceProviderExists(issuer);
     }
